DI the waypoint commands
